% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aRxiv}
\alias{aRxiv}
\title{Counts of the aRxiv field categories of 4 UChicago Statistics professors.
A dataset of compositional counts of 10 aRxiv field categories for 4 
professors' publications in aRxiv.}
\format{A matrix with 10 rows and 4 columns}
\usage{
aRxiv
}
\description{
Counts of the aRxiv field categories of 4 UChicago Statistics professors.
A dataset of compositional counts of 10 aRxiv field categories for 4 
professors' publications in aRxiv.
}
\keyword{datasets}
