% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commaletter.R
\name{commaletter}
\alias{commaletter}
\title{Logo plot for dot punctuation}
\usage{
commaletter(plot = FALSE, colfill = "green", y_pos = 0.1, x_pos = 0.5,
  lwd = 10)
}
\arguments{
\item{plot}{A binary. If FALSE, only outputs grid co-ordinates for the logo,
along with color labels. If TRUE, also plots the logo in a new grid
window. Defaults to FALSE.}

\item{colfill}{The color used to highlight the symbol.  Defaults to "green".}

\item{y_pos}{The y-position of the center of the comma in [0,1] X [0,1] window.
Defaults to 0.1.}

\item{x_pos}{The x-postition of the center of the comma in [0,1] X [0,1] window.
Defaults to 0.5.}

\item{lwd}{Specifies boundary of the comma. Defaults 10.}
}
\value{
Returns a list with the following items.
        \item{x}{X co-ordinates of the logo in the [0,1] X [0,1] grid window}
        \item{y}{Y co-ordinates of the logo in the [0,1] X [0,1] grid window}
        \item{id}{id vector representing blocks in the logo co-ordinates}
        \item{fill}{a vector equal to the number of distinct ids or blocks in
                   the logo, whose elements correspond to colors of these blocks}
}
\description{
Plots the symbol or logo for dot punctuation
}
\examples{
out <- commaletter(plot=TRUE)
}

