% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{consensus_PWM_codons}
\alias{consensus_PWM_codons}
\title{Consensus initiation contexts in seven different organisms. 
The consensus of initiation in different species was obtained from the Transterm
database. The data has been fetched from Ivanov et al 2012.}
\format{A list of matrices, each matrix with 4 rows and 15 columns}
\usage{
consensus_PWM_codons
}
\description{
Consensus initiation contexts in seven different organisms. 
The consensus of initiation in different species was obtained from the Transterm
database. The data has been fetched from Ivanov et al 2012.
}
\keyword{datasets}
