% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash.R
\name{dash}
\alias{dash}
\title{Dirichlet adaptive shrinkage of compositional data using dash}
\usage{
dash(comp_data, concentration = NULL, mode = NULL, weight = list(center =
  10, null = 1, corner = 1), def_positions = list(center = Inf, null = 1,
  corner = 0.5), optmethod = c("mixEM", "w_mixEM", "mixIP"),
  sample_weights = NULL, verbose = FALSE, fdr_bound = 50, bf = TRUE,
  pi_init = NULL, control = list(), reportcov = FALSE)
}
\arguments{
\item{comp_data, }{a n by m matrix where n represents the sample and m
represents the category of composition.}

\item{concentration}{a vector of concentration scales for different Dirichlet
compositions. Defaults to NULL, in which case, we append
concentration values of Inf, 100, 50, 20, 10, 5, 2, 1, 0.5
and 0.1.}

\item{mode}{An user defined mode/mean for the Dirichlet components. Defaults
to equal means for all components.}

\item{weight}{A list of three components which assigns weights to the center
Dirichlet component (determined by defined position of concentrations).
Defaults to center weight of 10 and null and corner weights of 1.}

\item{def_positions}{A list of three components which assigns concentration
values to center, null and corner components of Dirichlet
distributions. The default concentrations for center are Inf,
null are 1 and corner as 0.5.}

\item{optmethod}{The method for performing optimization of the mixture
proportions or grades of memberships for the different
Dirichlet compositions. Can be either of "mixEM", "w_mixEM"
or weighted mixEM and "mixIP" for interior point convex
optimization.}

\item{sample_weights}{The weights of the samples for performing the optimization.
Defaults to NULL, in which case the weight is same for each sample.}

\item{verbose}{if TRUE, outputs messages tracking progress of the method}

\item{fdr_bound}{The concentration bound used to detect false discoveries
of deviations from the mode.}

\item{bf}{A boolean (TRUE/FALSE) variable denoting whether log bayes factor
(with respect to category with smallest representation) is used in
optimization or the loglikelihood. Defaults to FALSE.}

\item{pi_init}{An initial starting value for the mixture proportions. Defaults
to same proportion for all categories.}

\item{control}{A list of control parameters for the SQUAREM/IP algorithm,
default value is set to be control.default=list(K = 1, method=3,
 square=TRUE, step.min0=1, step.max0=1, mstep=4, kr=1,
 objfn.inc=1,tol=1.e-07, maxiter=5000, trace=FALSE).}

\item{reportcov}{A boolean indicating whether the user wants to return
the covariance and correlation structure of the posterior.
Defaults to FALSE.}
}
\value{
A list, including the following,
        \code{fitted_pi}: The fitted values of mixture proportions for Dirichlet components
        \code{concentration}: The concentration scales of the Dirichlet compositions
        \code{prior}: Prior strengths of Dirichlet components
        \code{posterior_weights}: Posterior weights of each sample on each category posterior component
        \code{posmean}: Posterior means of compositional probability from dash fit of each sample
        \code{datamean}: Original compositional probability of each sample
        \code{poscov}: Posterior covariance structure for each sample (if \code{reportcov} TRUE)
        \code{poscor}: Posterior correlation structure for each sample (if \code{reportcov} TRUE)
        \code{center_prob_local}: Posterior probability on Inf concentration Dirichlet component
        \code{center_prob}: Posterior probability on Dirichlet components with concentration less than \code{fdr_bound}
        \code{corner_prob}: Posterior probability on Dirichlet components with concentration less than 1
}
\description{
Given a matrix of compositional counts data, with samples along
the rows and the categories of composition along columns, performs Bayesian
adaptive shrinkage of the compositions to produce refined composition probs.
}
\details{
The dash function provides a number of ways to perform
Empirical Bayes shrinkage estimation on compositional data (counts).

The inputs to dash is a matrix of compositional counts with samples along
rows and categories along columns. The method assumes that the compositional
counts data is generated from an underlying composition probability vector,
which follows a mixture of Dirichlet distributions centered at the
user defined mode (which defaults to means for all categories being equal).

We assume that the component Dirichlet distributions in the mixture have varying
degrees of concentration, varying from Inf (which is same as saying a point mass
at the mode), and then from high to low values of concentration and even concentration
values less than 1, which would represent spikes at the corners of the simplex.

The grades of memberships/ mixture proportions in different Dirichlet
components are estimated and post-hoc measures - posterior mean, posterior weights,
posterior center and corner probabilities etc are computed. The posterior mean
is considered as the shrunk compositional probability.
}
\examples{
mat <- rbind(c(5, 0, 2, 0),
             c(1, 1, 0, 1),
             c(100, 100, 50, 100),
             c(20, 50, 100, 10),
             c(10, 10, 200, 20),
             c(50, 54, 58, 53),
             c(1,1,1,3),
             c(2, 4, 1, 1))
out <- dash(xmat, optmethod = "mixEM", verbose=TRUE)
out <- dash(xmat, optmethod = "w_mixEM", verbose=TRUE)

}
