% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EBF1_disc1}
\alias{EBF1_disc1}
\title{EBF1_disc1 transcription factor position weight matrix
A dataset of composition probabilities of A, C, G and T in 10 positions of 
the motif.}
\format{A matrix with 4 rows and 10 columns}
\usage{
EBF1_disc1
}
\description{
EBF1_disc1 transcription factor position weight matrix
A dataset of composition probabilities of A, C, G and T in 10 positions of 
the motif.
}
\keyword{datasets}
