% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ELF1}
\alias{ELF1}
\title{ELF1 initiation contexts in metazoa, fungi and plants. 
The data has been fetched from Ivanov et al 2012.}
\format{A list of 3 matrices, each matrix with 4 rows and 15 columns}
\usage{
ELF1
}
\description{
ELF1 initiation contexts in metazoa, fungi and plants. 
The data has been fetched from Ivanov et al 2012.
}
\keyword{datasets}
