% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ELF1_PWM}
\alias{ELF1_PWM}
\title{Consensus and eIF1 initiation contexts in seven different species. 
The consensus of initiation in different species was obtained from the Transterm
database. The data has been fetched from Ivanov et al 2012.}
\format{A matrix with 4 rows and 15 columns}
\usage{
ELF1_PWM
}
\description{
Consensus and eIF1 initiation contexts in seven different species. 
The consensus of initiation in different species was obtained from the Transterm
database. The data has been fetched from Ivanov et al 2012.
}
\keyword{datasets}
