% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot.R
\name{get_plot}
\alias{get_plot}
\title{Logo plot for given co-ordinates and fill values}
\usage{
get_plot(x, y, id, fill, colfill, lwd = 10, fill_symbol = TRUE)
}
\arguments{
\item{x}{The X co-ordinates of the symbol}

\item{y}{The Y co-ordinates of the symbol}

\item{id}{The id of the different points based on how it will be grouped}

\item{fill}{The color vector used for filling the logo.}

\item{colfill}{The color label for the borders of the plot.}

\item{lwd}{Specifies the border width of the symbol. Defaults to 10.}

\item{fill_symbol}{A binary. If TRUE, the function would fill the symbol by
the color represented in \code{colfill}, else colors the boundary
of the symbol by \code{colfill}. Defaults to TRUE.}
}
\value{
Returns a plot of the symbol with the input co-ordinates and fill colors
}
\description{
Plots logo given the co-ordinates and fill values of the symbols.
}
\keyword{internal}

