% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_viewport_logo.R
\name{get_viewport_logo}
\alias{get_viewport_logo}
\title{Function for creating a multi-panel logo viewport}
\usage{
get_viewport_logo(layout.rows, layout.cols, widths.val = 6,
  heights.val = 20)
}
\arguments{
\item{layout.rows}{The number of rows in the panel}

\item{layout.cols}{The number of columns in the panel}

\item{widths_1}{The widths of each viewport in the panel}

\item{heights_1}{The heights of each viewport in the panel}
}
\value{
Creates a panel of viewports for each row and each column, to be
subsequently used for creating logo plots.
}
\description{
This is a void function to be run prior to a multi-panel Logolas
plot for a fixed number of rows and columns of the panel.
}
\examples{

get_viewport_logo(layout.rows = 2, layout.cols = 2)
}
