% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetConsensusSeq.R
\name{GetConsensusSeq}
\alias{GetConsensusSeq}
\title{Function for obtaining consensus sequence of DNA sequence
symbols from a PWM matrix}
\usage{
GetConsensusSeq(pwm)
}
\arguments{
\item{pwm}{The PWM table for a DNA sequence motif comprising of four bases, A, C
G, T.}
}
\value{
Returns the consensus sequence for the DNA sequence motif along the positions
using the Logolas nomenclature (highlighting both enrichment and depletion).
}
\description{
uses a special nomenclature (we call it the Logolas nomenclature)
to determine the consensus sequence of symbols based on the enrichment and
depletion of the symbols at each position. This approach is an alternative
to the getIUPAC() method used by the atSNP package.
}
\examples{

pwm=matrix(c(0.8,0.1,0.1,0,0.9,0.1,0,0,0.9,0.05,0.05,0,0.5,0.4,0,0.1,0.6,0.4,0,0,0.4,0.4,0.1,0.1,0.5,0,0.2,0.3,0.35,0.35,0.06,0.24,0.4,0.3,0.2,0.1,0.4,0.2,0.2,0.2,0.28,0.24,0.24,0.24,0.5,0.16,0.17,0.17,0.6,0.13,0.13,0.14,0.7,0.15,0.15,0),nrow = 4,byrow = F)
rownames(pwm)=c('A','C','G','T')
colnames(pwm)=1:ncol(pwm)
GetConsensusSeq(pwm)

}
