% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{himalayan_fauna}
\alias{himalayan_fauna}
\title{Proportional abundances of different bird species in the Himalayan mountains
A dataset of proportional composition of 140 bird species in 3 regions of 
the Himalayas.}
\format{A matrix with 140 rows and 3 columns}
\usage{
himalayan_fauna
}
\description{
Proportional abundances of different bird species in the Himalayan mountains
A dataset of proportional composition of 140 bird species in 3 regions of 
the Himalayas.
}
\keyword{datasets}
