% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{himalayan_fauna_2_clusters}
\alias{himalayan_fauna_2_clusters}
\title{Phylogenetic abundance of bird species families in two clusters of regions in
Himalayas}
\format{A matrix with bird species families along rows and the clusters
along columns}
\usage{
himalayan_fauna_2_clusters
}
\value{
A matrix with bird species families along rows and the clusters
along columns
}
\description{
Phylogenetic abundance of bird species families in two clusters of regions in
Himalayas
}
\keyword{datasets}
