% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Iletter.R
\name{Iletter}
\alias{Iletter}
\title{Logo plot for alphabet I}
\usage{
Iletter(plot = FALSE, fill_symbol = TRUE, colfill = "green", lwd = 10)
}
\arguments{
\item{plot}{A binary. If FALSE, only outputs grid co-ordinates for the logo,
along with color labels. If TRUE, also plots the logo in a new grid
window. Defaults to FALSE.}

\item{fill_symbol}{A binary. If TRUE, the function would fill the symbol by
the color represented in \Rcode{colfill}, else colors the boundary
of the symbol by \Rcode{colfill}. Defaults to TRUE.}

\item{colfill}{The color used to highlight the symbol.  Defaults to "green".}

\item{lwd}{Specifies the border width of the symbol. Defaults to 10.}
}
\value{
Returns a list with the following items.
        \item{x}{X co-ordinates of the logo in the [0,1] X [0,1] grid window}
        \item{y}{Y co-ordinates of the logo in the [0,1] X [0,1] grid window}
        \item{id}{id vector representing blocks in the logo co-ordinates}
        \item{fill}{a vector equal to the number of distinct ids or blocks in
                   the logo, whose elements correspond to colors of these blocks}
}
\description{
Plots the symbol or logo for alphabet I (uppercase).
}
\examples{
out <- Iletter(plot=TRUE)
}

