% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0letter.R, R/1letter.R, R/2letter.R,
%   R/3letter.R, R/4letter.R, R/5letter.R, R/6letter.R, R/7letter.R,
%   R/8letter.R, R/9letter.R, R/Aletter.R, R/Bletter.R, R/Cletter.R,
%   R/Dletter.R, R/Eletter.R, R/Fletter.R, R/Gletter.R, R/Hletter.R,
%   R/Iletter.R, R/Jletter.R, R/Kletter.R, R/Lletter.R, R/Mletter.R,
%   R/Nletter.R, R/Oletter.R, R/Pletter.R, R/Qletter.R, R/Rletter.R,
%   R/Sletter.R, R/Tletter.R, R/Uletter.R, R/Vletter.R, R/Wletter.R,
%   R/Xletter.R, R/Yletter.R, R/Zletter.R, R/colonletter.R, R/commaletter.R,
%   R/dashletter.R, R/dotletter.R, R/leftarrowletter.R, R/rightarrowletter.R,
%   R/semicolonletter.R
\name{letters}
\alias{letters}
\alias{zeroletter}
\alias{oneletter}
\alias{twoletter}
\alias{threeletter}
\alias{fourletter}
\alias{fiveletter}
\alias{sixletter}
\alias{sevenletter}
\alias{eightletter}
\alias{nineletter}
\alias{Aletter}
\alias{Bletter}
\alias{Cletter}
\alias{Dletter}
\alias{Eletter}
\alias{Fletter}
\alias{Gletter}
\alias{Hletter}
\alias{Iletter}
\alias{Jletter}
\alias{Kletter}
\alias{Lletter}
\alias{Mletter}
\alias{Nletter}
\alias{Oletter}
\alias{Pletter}
\alias{Qletter}
\alias{Rletter}
\alias{Sletter}
\alias{Tletter}
\alias{Uletter}
\alias{Vletter}
\alias{Wletter}
\alias{Xletter}
\alias{Yletter}
\alias{Zletter}
\alias{colonletter}
\alias{commaletter}
\alias{dashletter}
\alias{dotletter}
\alias{leftarrowletter}
\alias{rightarrowletter}
\alias{semicolonletter}
\title{Logo plot for numbers 0-9, A-Z and punctuation marks}
\usage{
zeroletter(plot = FALSE, fill_symbol = TRUE, colfill = "green",
  lwd = 10)

oneletter(plot = FALSE, fill_symbol = TRUE, colfill = "green", lwd = 10)

twoletter(plot = FALSE, fill_symbol = TRUE, colfill = "green", lwd = 10)

threeletter(plot = FALSE, fill_symbol = TRUE, colfill = "green",
  lwd = 10)

fourletter(plot = FALSE, fill_symbol = TRUE, colfill = "green",
  lwd = 10)

fiveletter(plot = FALSE, fill_symbol = TRUE, colfill = "green",
  lwd = 10)

sixletter(plot = FALSE, fill_symbol = TRUE, colfill = "green", lwd = 10)

sevenletter(plot = FALSE, fill_symbol = TRUE, colfill = "green",
  lwd = 10)

eightletter(plot = FALSE, fill_symbol = TRUE, colfill = "green",
  lwd = 10)

nineletter(plot = FALSE, fill_symbol = TRUE, colfill = "green",
  lwd = 10)

Aletter(plot = FALSE, fill_symbol = TRUE, colfill = "green", lwd = 10)

Bletter(plot = FALSE, fill_symbol = TRUE, colfill = "green", lwd = 10)

Cletter(plot = FALSE, fill_symbol = TRUE, colfill = "green", lwd = 10)

Dletter(plot = FALSE, fill_symbol = TRUE, colfill = "green", lwd = 10)

Eletter(plot = FALSE, fill_symbol = TRUE, colfill = "green", lwd = 10)

Fletter(plot = FALSE, fill_symbol = TRUE, colfill = "green", lwd = 10)

Gletter(plot = FALSE, fill_symbol = TRUE, colfill = "green", lwd = 10)

Hletter(plot = FALSE, fill_symbol = TRUE, colfill = "green", lwd = 10)

Iletter(plot = FALSE, fill_symbol = TRUE, colfill = "green", lwd = 10)

Jletter(plot = FALSE, fill_symbol = TRUE, colfill = "green", lwd = 10)

Kletter(plot = FALSE, fill_symbol = TRUE, colfill = "green", lwd = 10)

Lletter(plot = FALSE, fill_symbol = TRUE, colfill = "green", lwd = 10)

Mletter(plot = FALSE, fill_symbol = TRUE, colfill = "green", lwd = 10)

Nletter(plot = FALSE, fill_symbol = TRUE, colfill = "green", lwd = 10)

Oletter(plot = FALSE, fill_symbol = TRUE, colfill = "green", lwd = 10)

Pletter(plot = FALSE, fill_symbol = TRUE, colfill = "green", lwd = 10)

Qletter(plot = FALSE, fill_symbol = TRUE, colfill = "green", lwd = 10)

Rletter(plot = FALSE, fill_symbol = TRUE, colfill = "green", lwd = 10)

Sletter(plot = FALSE, fill_symbol = TRUE, colfill = "green", lwd = 10)

Tletter(plot = FALSE, fill_symbol = TRUE, colfill = "green", lwd = 10)

Uletter(plot = FALSE, fill_symbol = TRUE, colfill = "green", lwd = 10)

Vletter(plot = FALSE, fill_symbol = TRUE, colfill = "green", lwd = 10)

Wletter(plot = FALSE, fill_symbol = TRUE, colfill = "green", lwd = 10)

Xletter(plot = FALSE, fill_symbol = TRUE, colfill = "green", lwd = 10)

Yletter(plot = FALSE, fill_symbol = TRUE, colfill = "green", lwd = 10)

Zletter(plot = FALSE, fill_symbol = TRUE, colfill = "green", lwd = 10)

colonletter(plot = FALSE, colfill = "green", y_pos_1 = 0.3,
  y_pos_2 = 0.7, x_pos = 0.5, rad = 0.1, lwd = 10)

commaletter(plot = FALSE, colfill = "green", y_pos = 0.1, x_pos = 0.5,
  lwd = 10)

dashletter(plot = FALSE, fill_symbol = TRUE, colfill = "green",
  lwd = 10)

dotletter(plot = FALSE, colfill = "green", y_pos = 0.1, x_pos = 0.5,
  rad = 0.1)

leftarrowletter(plot = FALSE, fill_symbol = TRUE, colfill = "green",
  lwd = 10)

rightarrowletter(plot = FALSE, fill_symbol = TRUE, colfill = "green",
  lwd = 10)

semicolonletter(plot = FALSE, colfill = "green", y_pos_1 = 0.1,
  y_pos_2 = 0.4, x_pos = 0.5, rad = 0.1, lwd = 10)
}
\arguments{
\item{plot}{A binary. If FALSE, only outputs grid co-ordinates for the logo,
along with color labels. If TRUE, also plots the logo in a new grid
window. Defaults to FALSE.}

\item{fill_symbol}{A binary. If TRUE, the function would fill the symbol by
the color represented in \code{colfill}, else colors the boundary
of the symbol by \code{colfill}. Defaults to TRUE.}

\item{colfill}{The color used to highlight the symbol.  Defaults to "green".}

\item{lwd}{Specifies the border width of the symbol. Defaults to 10.}

\item{y_pos_1}{The y-position of the center of the lower dot in [0,1] X [0,1]
window. Defaults to 0.1.}

\item{y_pos_2}{The y-position of the center of the upper dot in [0,1] X [0,1] window.
Defaults to 0.7.}

\item{x_pos}{The x-postition of the center of the dot in [0,1] X [0,1] window.
Defaults to 0.5.}

\item{rad}{The radius of the indivudual symbols in the semicolon.}

\item{y_pos}{The y-position of the center of the comma in [0,1] X [0,1] window.
Defaults to 0.1.}

\item{lwd}{Specifies the border width of the symbol. Defaults to 10.}

\item{x_pos}{The x-postition of the center of the comma in [0,1] X [0,1] window.
Defaults to 0.5.}

\item{lwd}{Specifies boundary of the comma. Defaults 10.}

\item{y_pos}{The y-position of the center of the dot in [0,1] X [0,1] window.
Defaults to 0.1.}

\item{x_pos}{The x-postition of the center of the dot in [0,1] X [0,1] window.
Defaults to 0.5.}

\item{rad}{The radius of the dot. Defaults to 0.1.}

\item{y_pos_1}{The y-position of the center of the lower dot in [0,1] X [0,1]
window. Defaults to 0.1.}

\item{y_pos_2}{The y-position of the center of the upper dot in [0,1] X [0,1] window.
Defaults to 0.7.}

\item{x_pos}{The x-postition of the center of the dot in [0,1] X [0,1] window.
Defaults to 0.5.}

\item{rad}{The radius of the indivudual symbols in the semicolon.}

\item{lwd}{Specifies the border width of the symbol. Defaults to 10.}
}
\value{
Returns a list with the following items.
        \item{x}{X co-ordinates of the logo in the [0,1] X [0,1] grid window}
        \item{y}{Y co-ordinates of the logo in the [0,1] X [0,1] grid window}
        \item{id}{id vector representing blocks in the logo co-ordinates}
        \item{fill}{a vector equal to the number of distinct ids or blocks in
                   the logo, whose elements correspond to colors of these blocks}
}
\description{
Plots the symbol or logo for numbers 0-9 and alphabets A-Z,
left arrow, right arrow, colon, semic colon, dash, comma and dot.

Plots the symbol or logo for a colon.

Plots the symbol or logo for comma punctuation

Plots the symbol or logo for dot punctuation

Plots the symbol or logo for a semi-colon.
}
\examples{
out <- zeroletter(plot=TRUE)
out <- oneletter(plot=TRUE)
out <- twoletter(plot=TRUE)
out <- threeletter(plot=TRUE)
out <- fourletter(plot=TRUE)
out <- fiveletter(plot=TRUE)
out <- sixletter(plot=TRUE)
out <- sevenletter(plot=TRUE)
out <- eightletter(plot=TRUE)
out <- nineletter(plot=TRUE)
out <- Aletter(plot=TRUE)
out <- Bletter(plot=TRUE)

out <- Cletter(plot=TRUE)
out <- Dletter(plot=TRUE)

out <- Eletter(plot=TRUE)
out <- Fletter(plot=TRUE)
out <- Gletter(plot=TRUE)
out <- Hletter(plot=TRUE)
out <- Iletter(plot=TRUE)
out <- Jletter(plot=TRUE)
out <- Kletter(plot=TRUE)
out <- Lletter(plot=TRUE)
out <- Mletter(plot=TRUE)
out <- Nletter(plot=TRUE)
out <- Oletter(plot=TRUE)
out <- Pletter(plot=TRUE)
out <- Qletter(plot=TRUE)
out <- Rletter(plot=TRUE)
out <- Sletter(plot=TRUE)
out <- Tletter(plot=TRUE)
out <- Uletter(plot=TRUE)
out <- Vletter(plot=TRUE)
out <- Wletter(plot=TRUE)
out <- Xletter(plot=TRUE)
out <- Yletter(plot=TRUE)
out <- Zletter(plot=TRUE)
out <- colonletter(plot=TRUE)
out <- commaletter(plot=TRUE)
out <- dashletter(plot=TRUE)
out <- dotletter(plot=TRUE)
out <- leftarrowletter(plot=TRUE)
out <- rightarrowletter(plot=TRUE)
out <- semicolonletter(plot=TRUE)
}
\keyword{internal}
