% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logomaker_format_input.R
\name{logomaker_format_input}
\alias{logomaker_format_input}
\title{Format conversion from FASTA, Genbank, CLustalW. PhyLIP formats to 
vector of aligned sequences for logomaker}
\usage{
logomaker_format_input(filename = NULL, accession = NULL, lower_base = 1,
  upper_base = Inf)
}
\arguments{
\item{filename}{Name of the file with stored aligned sequences - may be of
the format .phylip (PHYLIP), .aln (CLUSTAL) or .fasta(FASTA).
May also be NULL if Genbank accession codes input
\code{accession} is non NULL}

\item{accession}{A vector of accession numbers for GenBank. Not null when
\code{filename} is NULL.}

\item{lower_base}{The base position from which to start recording the 
sequence. Defaults to 1.}

\item{upper_base}{The base position at which to end recording the sequence.
Defaults to Inf. That would comprise of the full sequence
if \code{lower_base = 1}}
}
\value{
Returns a vector of aligned sequences, that are ready to be fed as 
        input to the \code{logomaker} function
}
\description{
Takes as input a FASTA/ClustalW/PhyLIP files or accession numbers
from Genbank and converts the sequences into a vector of aligned sequences, 
ready to be fed into the \code{logomaker} function. 
(see examples in \code{logomaker} to see its input format, which is same as 
the output format for this function).
}
\examples{

library(seqinr)
out <- logomaker_format_input(filename = system.file("sequences/gopher.fasta", 
                              package = "seqinr"),
                              lower_base = 1, upper_base = 15)
out <- logomaker_format_input(filename = system.file("sequences/test.aln", 
                              package = "seqinr"),
                              lower_base = 1, upper_base = 15)
out <- logomaker_format_input(filename = system.file("sequences/test.phylip", 
                              package = "seqinr"),
                              lower_base = 1, upper_base = 15)
out <- logomaker_format_input(accession =  c("U15717", "U15718", "U15719",
                          "U15720", "U15721", "U15722", "U15723", "U15724"),
                           lower_base = 1, upper_base = 10)

}
