% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mutation_sig}
\alias{mutation_sig}
\title{Compositional mutational signature data, with mismatch and flanking base
frequencies reported
A dataset of compositional weights for mismatches in the "0"the position 
(center of signature) and
that of the bases A, C, G, T for two left flanking bases (-1, -2) and two 
right flanking bases (1, 2). The data here corresponds to the 12th 
mutational signature profile from Shiraishi et al 2015 paper.}
\format{A matrix with 10 rows (4 bases and 6 types of mismatches) and 
5 columns (center - 0, flanking bases
1,2 on the left and flanking bases 1, 2 on the right of the center)}
\usage{
mutation_sig
}
\description{
Compositional mutational signature data, with mismatch and flanking base
frequencies reported
A dataset of compositional weights for mismatches in the "0"the position 
(center of signature) and
that of the bases A, C, G, T for two left flanking bases (-1, -2) and two 
right flanking bases (1, 2). The data here corresponds to the 12th 
mutational signature profile from Shiraishi et al 2015 paper.
}
\keyword{datasets}
