% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{N_Glycosyl_sequences}
\alias{N_Glycosyl_sequences}
\title{Amino acid sequences from UniProt database that constitute N-glycosylation sites.
The data has been fetched from UniProt website.}
\format{A vector of sequences, with the glycosylation site at the center (usually N),
and 5 bases on either side of it.}
\usage{
N_Glycosyl_sequences
}
\description{
Amino acid sequences from UniProt database that constitute N-glycosylation sites.
The data has been fetched from UniProt website.
}
\keyword{datasets}
