% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neg_ic_computer.R
\name{neg_ic_computer}
\alias{neg_ic_computer}
\title{Negative scaling output for nlogomaker function}
\usage{
neg_ic_computer(table, hist = FALSE, alpha = 1)
}
\arguments{
\item{table}{The input table (data frame or matrix) of counts across different
logos or symbols (specified along the rows) ans across different sites or
positions or groups (specified along the columns).}

\item{hist}{Whether to use the hist method or the information criterion
method to determine the heights of the logos.}

\item{alpha}{The Renyi entropy tuning parameter which is used in case of
scaling of the bar heights by information criterion. The default tuning
parameter value is 1, which corresponds to Shannon entropy.}
}
\value{
Returns the enrichment heights in the positive and the negative scale.
  \code{pos}: scaling on the positive scale
  \code{neg}: scaling on the negative scale
  \code{scales}: the scaling factor for each site positive and negative heights
  \code{ic_pos}: height of the positive bar
  \code{ic_neg}: height of the negative bar
}
\description{
a scaling function for \code{nlogomaker()} to separately scale the
positive and the negative axes heights in the enrichment logo plots.
}
\examples{

m = matrix(rep(0,48),4,12)
m[1,] = c(0,0,2.5,7,0,0,0,0,0,0,1,0)
m[2,] = c(4,6,3,1,0,0,0,0,0,5,0,5)
m[3,] = c(0,0,0,0,0,1,8,0,0,1,1,2)
m[4,] = c(4,2,2.5,0,8,7,0,8,8,2,6,1)
rownames(m) = c("A", "C", "G", "T")
colnames(m) = 1:12
m=m/8
ic2 <- neg_ic_computer(m)

ic2$ic_pos
ic2$ic_neg
ic2$scales
}
