% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlogomaker.R
\name{nlogomaker}
\alias{nlogomaker}
\title{Main workhorse function that builds negative logo plots}
\usage{
nlogomaker(table = NULL, ic = FALSE, score = c("diff", "log",
  "log-odds", "probKL", "ratio", "unscaled_log", "wKL", "preclog"),
  color_profile, total_chars = c("A", "B", "C", "D", "E", "F", "G", "H",
  "I", "J", "K", "L", "M", "N", "O", "P", "Q", "R", "S", "T", "U", "V",
  "W", "X", "Y", "Z", "zero", "one", "two", "three", "four", "five", "six",
  "seven", "eight", "nine", "dot", "comma", "dash", "colon", "semicolon",
  "leftarrow", "rightarrow"), bg = NULL, llambda = NULL, tol = 0,
  frame_width = NULL, yscale_change = TRUE, pop_name = NULL,
  addlogos = NULL, addlogos_text = NULL, newpage = TRUE,
  yrange = NULL, xaxis = TRUE, yaxis = TRUE, xaxis_fontsize = 10,
  xlab_fontsize = 15, y_fontsize = 15, main_fontsize = 16,
  start = 0.001, xlab = "X", ylab = "Enrichment Score",
  col_line_split = "grey80", control = list())
}
\arguments{
\item{table}{The input table (data frame or matrix) of counts
across different logos or symbols (specified along the rows) 
and across different sites or positions or groups (specified 
along the columns).}

\item{ic}{Boolean, denoting whether information content based 
scaling is used on top of the scoring scheme used or not. 
Default is FALSE}

\item{score}{Can take either of the options - \code{diff},
\code{log}, \code{log-odds}, \code{probKL}, \code{ratio}, 
\code{unscaled_log}, \code{wKL}. Each option corresponds to 
 a different scoring scheme. The most recommended option is 
 \code{log}.}

\item{color_profile}{A list containing two elements - "type" and "col".
 The type can be of three types - "per-row", "per-column" and 
 "per-symbol". The "col" element is a vector of colors, of same length
 as number of rows in table for "per-row" (assigning a color to each 
 string), of same length as number of columns in table for "per-column"
(assuming a color for each column), or a distinct color for a distinct
 symbol in "per-symbol". For "per-symbol", the length of the 
 \code{color_profile$col} should be same as library size of the logos, 
 but if the vector of colors provided is more or less, we can 
 downsample or upsample the colors as required. The colors are 
 matched with the symbols in the \code{total_chars}}

\item{total_chars}{The total number of character symbols in the user
library. The default is the default library provided by Logolas,
but the user can add symbols that he creates to this list.}

\item{bg}{The background probability, which defaults to NULL, 
in which case equal probability is assigned to each symbol. The user 
can however specify a vector (equal to in length to the number of 
symbols) which specifies the background probability for each symbol 
and assumes this background probability to be the same across the 
columns (sites), or a matrix, whose each cell specifies
the background probability of the symbols for each position.}

\item{llambda}{The log lambda matrix computed after stabilization step (if
used in \code{logomaker} function). Used for computing the
heights when not NULL.}

\item{tol}{The tolerance for the KL-divergence of the positional weight 
data and background probabilities (for preclog).}

\item{frame_width}{The width of the frames for individual 
site/postion/column in the logo plot. As default, all the 
columns have same width, equal to 1.}

\item{yscale_change}{If TRUE, adjusts the Y axis scale based on
the size of the bars, else keeps it to the maximum value possible,
which is \code{ceiling(max(ic)} under \code{ic_computer} 
defined IC criteria.}

\item{pop_name}{User can mention a name of the population for which
the logo plot is created. Defaults to NULL when no population name 
is mentioned.}

\item{addlogos}{Vector of additional logos/symbols defined by user}

\item{addlogos_text}{Vector of the names given to the additional 
logos/symbols defined by user.}

\item{newpage}{if TRUE, plots the logo plot in a new page. Defaults 
to TRUE.}

\item{yrange}{The limit of the Y axis.}

\item{xaxis}{Binary specifying if there should be a X axis in 
the logo plot or not. Defaults to TRUE.}

\item{yaxis}{Binary specifying if there should be a Y axis in 
the logo plot or not. Defaults to TRUE.}

\item{xaxis_fontsize}{The size of the X-axis axis ticks.}

\item{xlab_fontsize}{The size of the X-axis label.}

\item{y_fontsize}{The size of the Y-axis font.}

\item{main_fontsize}{The size of the title.}

\item{start}{The starting point in Y axis for the first logo. 
Default is 0.0001 which is very close to 0.}

\item{xlab}{X axis label}

\item{ylab}{Y axis label}

\item{col_line_split}{The color of the line split between the 
consecutive groups or blocks}

\item{control}{control parameters fixing whether the height of the 
logos is detrmined by IC or histogram proportions (\code{hist}),
the scales for the plot (\code{scale0}, \code{scale1}), the quantile 
adjustment (\code{quant}), whether to use symmetric KL for scaling 
(\code{symm}), whether the symbols should be filled with color or 
border colored (\code{tofill_pos, tofill_neg}), the Renyi 
alpha parameter for the entropy calculation (\code{alpha}), the gap between
ylabel and y-axis and xlabel and x-axis texts (\code{gap_ylab}, \code{gap_xlab}),
the viewport configuration details for the plot 
(\code{viewport.margin.bottom},\code{viewport.margin.left},
 \code{viewport.margin.top}, \code{viewport.margin.right}),
 whether the height of the logos would be fixed
apriori or determined by the PWM matrix as in seqLogo
(\code{use_seqLogo_heights}) etc.}
}
\value{
Plots the logo plot for the table data, with column names 
representing the sites/blocks and the row names denoting the symbols
for which logos are plotted
}
\description{
stacks logos created by the \code{makemylogo} function
 on top of each other to build the logo plot.
}
