% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudocount_adjust.R
\name{pseudocount_adjust}
\alias{pseudocount_adjust}
\title{PFM/PWM Pseudo-count adjustment to get rid of 0 entries.}
\usage{
pseudocount_adjust(table, pseudocount = NULL)
}
\arguments{
\item{table}{A table (PFM/PWM) with the symbols along the rows and the logo stack
positions along the columns, possibly containing 0 values.}

\item{pseudocount}{A small pseudocount to be added mainly to bypass 0 entries. 
Default is NULL. If \code{table} is a counts matrix, 
the default changes to 0.5, if \code{table} is a 
positional weight matrix, the default becomes 0.001 times
the minimum non-zero value of the table.}
}
\value{
Returns a new table that pseudo-count adjusts the 0 values.
}
\description{
Building a pseudo-count adjusted matrix from a PWM
(positional weight matrix) or a PFM (positional frequency matrix) with 0 entries.
}
\examples{
data("mutation_sig")
pseudocount_adjust(mutation_sig)
}
