% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semicolonletter.R
\name{semicolonletter}
\alias{semicolonletter}
\title{Logo plot for a semi-colon}
\usage{
semicolonletter(plot = FALSE, colfill = "green", y_pos_1 = 0.1,
  y_pos_2 = 0.4, x_pos = 0.5, rad = 0.1, lwd = 10)
}
\arguments{
\item{plot}{A binary. If FALSE, only outputs grid co-ordinates for the logo,
along with color labels. If TRUE, also plots the logo in a new grid
window. Defaults to FALSE.}

\item{colfill}{The color used to highlight the symbol.  Defaults to "green".}

\item{y_pos_1}{The y-position of the center of the lower dot in [0,1] X [0,1]
window. Defaults to 0.1.}

\item{y_pos_2}{The y-position of the center of the dot in [0,1] X [0,1] window.
Defaults to 0.7.}

\item{x_pos}{The x-postition of the center of the dot in [0,1] X [0,1] window.
Defaults to 0.5.}

\item{rad}{The radius of the indivudual symbols in the semicolon.}

\item{lwd}{Specifies the border width of the symbol. Defaults to 10.}

\item{fill_symbol}{A binary. If TRUE, the function would fill the symbol by
the color represented in \Rcode{colfill}, else colors the boundary
of the symbol by \Rcode{colfill}. Defaults to TRUE.}

\item{y_pos_1}{The y-position of the center of the dot in [0,1] X [0,1] window.
Defaults to 0.1.}
}
\value{
Returns a list with the following items.
        \item{x}{X co-ordinates of the logo in the [0,1] X [0,1] grid window}
        \item{y}{Y co-ordinates of the logo in the [0,1] X [0,1] grid window}
        \item{id}{id vector representing blocks in the logo co-ordinates}
        \item{fill}{a vector equal to the number of distinct ids or blocks in
                   the logo, whose elements correspond to colors of these blocks}
}
\description{
Plots the symbol or logo for a semi-colon.
}
\examples{
out <- semicolonletter(plot=TRUE)
}

