% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UMI}
\alias{UMI}
\title{An example of the position weight matrix of the bases in the first 30 
positions from ends of sequenced
reads (together with the barcode)
A dataset of composition probabilities of A, C, G and T in 30 positions 
from the 5' end of reads.}
\format{A matrix with 4 rows and 30 columns}
\usage{
UMI
}
\description{
An example of the position weight matrix of the bases in the first 30 
positions from ends of sequenced
reads (together with the barcode)
A dataset of composition probabilities of A, C, G and T in 30 positions 
from the 5' end of reads.
}
\keyword{datasets}
