% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makemylogo.R
\name{makemylogo}
\alias{makemylogo}
\title{Logo maker for a given English alphanumeric with common punctuations}
\usage{
makemylogo(name, colfill = "orange", plot = FALSE, total_chars = c("A",
  "B", "C", "D", "E", "F", "G", "H", "I", "J", "K", "L", "M", "N", "O", "P",
  "Q", "R", "S", "T", "U", "V", "W", "X", "Y", "Z", "zero", "one", "two",
  "three", "four", "five", "six", "seven", "eight", "nine", "dot", "comma",
  "dash", "colon", "semicolon", "leftarrow", "rightarrow"), addlogos = NULL,
  addlogos_text = NULL)
}
\arguments{
\item{name}{A English name, or alphanumeric, containing English alphabets,
numbers, dots, dashes, arroww, colons, semicolons, comma among punctuations.}

\item{colfill}{The color used for the symbol}

\item{plot}{binary, if FALSE, returns only the co-ordinates of the symbol in the
[0,1] X [0,1] grid, along with block id labels and their corresponding colors.
If TRUE, plots the symbol with specified color in a new grid window.}

\item{total_chars}{The total number of character symbols in the user library. The default
is the default library provided by Logolas, but the user can add symbols that he creates
to this list.}

\item{addlogos}{Vector of additional logos/symbols defined by user}

\item{addlogos_text}{Vector of the names given to the additional
logos/symbols defined by user.}
}
\value{
Along with symbol plot, if plot is TRUE, returns a list with the
following items.
        \item{x}{X co-ordinates of the logo in the [0,1] X [0,1] grid window}
        \item{y}{Y co-ordinates of the logo in the [0,1] X [0,1] grid window}
        \item{id}{id vector representing blocks in the logo co-ordinates}
        \item{fill}{a vector equal to the number of distinct ids or blocks in
                   the logo, whose elements correspond to colors of these
                   blocks}
}
\description{
Plots logo for a given english symbol or name that contains
English alphabets, numbers or punctuations like dots, dashes, etc. This is
the skeleton used by the \code{logomaker} function of the package to create
distinct logos for distinct alphanumeric symbols.
}
\examples{

makemylogo("KUSHAL")
cols = RColorBrewer::brewer.pal.info[RColorBrewer::brewer.pal.info$category == 'qual',]
col_vector = unlist(mapply(RColorBrewer::brewer.pal, cols$maxcolors, rownames(cols)))
makemylogo("Evening", plot=TRUE, colfill=col_vector)
}
