% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_augment.R
\name{zero_augment}
\alias{zero_augment}
\title{PFM/PWM Table Zero Augmentation}
\usage{
zero_augment(table, eps = 1e-10)
}
\arguments{
\item{table}{A table (PFM/PWM) with the symbols along the rows and the logo stack
positions along the columns, possibly containing 0 values.}
}
\value{
Returns a new table that augments the zeros in the input \code{table}.
}
\description{
Building the zero-augmented positional weight matrix from a PWM
(positional weight matrix) or a PFM (positional frequency matrix) with 0 entries.
}
\examples{
data("mutation_sig")
zero_augment(mutation_sig)
}
